port rpc tests from python2 to python3
This commit includes code from:

https://github.com/mdr0id/zcash/tree/rpc_py2_to_py3

In particular the imlementation of test_framework/mininode.py is derived
from:

https://github.com/mdr0id/zcash/blob/2e0345f4843d3504ee4eed9597ab6ae2a53767f7/qa/rpc-tests/test_framework/mininode.py

As of this commit 4 tests (not including the "-extended" set fail
intermittently:

  * wallet_shieldingcoinbase.py
  * wallet_listreceived.py
  * wallet_listnotes.py
  * sprout_sapling_migration.py

Each of these tests sometimes pass when run individually.
At least "sprout_sapling_migration" and, If I Recall Correctly (IIRC),
each of the other ones also is more likely to pass if the evaluation
of certain assertions is _delayed_.   My hypothesis is that there is a
delay between the execution of an rpc command, and the propagation
of the resultant state to all components of the system under test.
It would be interesting to see if these tests also intermittently fail
under the python2 version of the rpc tests.

My test environment has generally been a python3-only environment in
order to eliminate the possibility of an undetected python2 dependency.

The "-extended" tests have not been shown to pass, and may contain
obsolete python2 idioms.

Decimal encoding during JSON parsing is handled by the jsonutil
dependency.

It is my hope that bespoke one-off code (e.g.
test_framework.utils.assert_equal) will be replaced with standard
libraries in the near future (e.g. unittest.TestCase:
(@daira suggested this somewhere but I've lost the URL)) thereby:

  * reducing the size of the RPC codebase
  * increasing the readability of the codebase (by standardizing)
  * decreasing the surface area for potential bugs, etc.